Revert "Merge remote-tracking branch 'remotes/origin/CalebFlick' into Programming"
This reverts commit 88ec5bd564aff9146a92fa4dec19e1f69c5484a5, reversing
changes made to dbd9966cccbe1391104e4be3c8c86ece7f1251aa.

+++ Assets/Scripts/Combat/DamagePlayer.cs
@@ -9,7 +9,7 @@ using UnityEngine;
 public class DamagePlayer : MonoBehaviour
 {
     [SerializeField]
-    private int attackDamage = 10;
+    private float attackDamage = 10.0f;
 
     private void OnTriggerEnter(Collider other)
     {
